+++ CHANGELOG.rst
@@ -8,6 +8,7 @@
 * Properly sends session ID on first request without needing to call session()
 * Search and Match return a single object instead of a one-item list
 * Allows authorize() to supply arbitrary parameters
+* The login.py example reads the password more portably
 
 
 0.1 (21 Jan 2011)